Extract caching logic from `BaseHMPPSClient`
This commit performs preparatory work to resolve the bug described in
[#1570](https://trello.com/c/KU3pJJox/1570-redis-cache-keys-are-required-for-rendering-the-assessment-page).
It refactors the specifics of interacting with the cache out of
`BaseHMPPSClient.doRequest` into several methods on a new
`WebClientCache` class:
- `checkPreemptiveCacheStatus`
- `tryGetCachedValue`
- `cacheSuccessfulWebClientResponse`
- `cacheFailedWebClientResponse`

This provides a more obvious separation of concerns and will make it
easier to rework the logic to avoid throwing the exception causing the
bug.
