+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/client/BaseHMPPSClient.kt
@@ -1,25 +1,19 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.client
 
-import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
-import io.sentry.Sentry
-import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpMethod
 import org.springframework.http.HttpStatus
 import org.springframework.web.reactive.function.client.WebClient
 import org.springframework.web.reactive.function.client.WebClientResponseException
-import java.time.Duration
-import java.time.Instant
+import java.util.concurrent.atomic.AtomicInteger
 
 abstract class BaseHMPPSClient(
   private val webClient: WebClient,
   private val objectMapper: ObjectMapper,
-  private val redisTemplate: RedisTemplate<String, String>,
-  private val preemptiveCacheKeyPrefix: String,
+  private val webClientCache: WebClientCache,
 ) {
   protected inline fun <reified ResponseType : Any> getRequest(noinline requestBuilderConfiguration: HMPPSRequestConfiguration.() -> Unit): ClientResult<ResponseType> =
     request(HttpMethod.GET, requestBuilderConfiguration)
@@ -36,60 +30,27 @@ abstract class BaseHMPPSClient(
   protected inline fun <reified ResponseType : Any> patchRequest(noinline requestBuilderConfiguration: HMPPSRequestConfiguration.() -> Unit): ClientResult<ResponseType> =
     request(HttpMethod.PATCH, requestBuilderConfiguration)
 
-  protected fun checkPreemptiveCacheStatus(cacheConfig: PreemptiveCacheConfig, key: String): PreemptiveCacheEntryStatus {
-    val cacheKeySet = CacheKeySet(preemptiveCacheKeyPrefix, cacheConfig.cacheName, key)
-
-    val cacheEntryMetadata = getCacheEntryMetadataIfExists(cacheKeySet.metadataKey)
-
-    return when {
-      cacheEntryMetadata == null -> PreemptiveCacheEntryStatus.MISS
-      cacheEntryMetadata.refreshableAfter < Instant.now() -> PreemptiveCacheEntryStatus.REQUIRES_REFRESH
-      else -> PreemptiveCacheEntryStatus.EXISTS
-    }
-  }
+  protected fun checkPreemptiveCacheStatus(cacheConfig: WebClientCache.PreemptiveCacheConfig, key: String): PreemptiveCacheEntryStatus =
+    webClientCache.checkPreemptiveCacheStatus(cacheConfig, key)
 
   protected inline fun <reified ResponseType : Any> request(method: HttpMethod, noinline requestBuilderConfiguration: HMPPSRequestConfiguration.() -> Unit): ClientResult<ResponseType> {
-    val clazz = if (ResponseType::class.java.typeParameters.any()) null else ResponseType::class.java
-    val typeReference = if (ResponseType::class.java.typeParameters.any()) object : TypeReference<ResponseType>() {} else null
+    val typeReference = object : TypeReference<ResponseType>() {}
 
-    return doRequest(clazz, typeReference, method, requestBuilderConfiguration)
+    return doRequest(typeReference, method, requestBuilderConfiguration)
   }
 
-  fun <ResponseType : Any> doRequest(clazz: Class<ResponseType>?, typeReference: TypeReference<ResponseType>?, method: HttpMethod, requestBuilderConfiguration: HMPPSRequestConfiguration.() -> Unit): ClientResult<ResponseType> {
+  fun <ResponseType : Any> doRequest(typeReference: TypeReference<ResponseType>, method: HttpMethod, requestBuilderConfiguration: HMPPSRequestConfiguration.() -> Unit): ClientResult<ResponseType> {
     val requestBuilder = HMPPSRequestConfiguration()
     requestBuilderConfiguration(requestBuilder)
 
     val cacheConfig = requestBuilder.preemptiveCacheConfig
 
-    var attempt = 1
+    val attempt = AtomicInteger(1)
 
     try {
       if (cacheConfig != null) {
-        val cacheKeySet = getCacheKeySet(requestBuilder, cacheConfig)
-
-        if (!requestBuilder.isPreemptiveCall) {
-          val pollingStart = System.currentTimeMillis()
-
-          do {
-            val cacheEntryMetadata = getCacheEntryMetadataIfExists(cacheKeySet.metadataKey)
-
-            if (cacheEntryMetadata == null) {
-              Thread.sleep(500)
-              continue
-            }
-
-            return resultFromCacheMetadata(cacheEntryMetadata, cacheKeySet, typeReference, clazz)
-          } while (System.currentTimeMillis() - pollingStart < requestBuilder.preemptiveCacheTimeoutMs)
-
-          return ClientResult.Failure.PreemptiveCacheTimeout(cacheConfig.cacheName, cacheKeySet.metadataKey, requestBuilder.preemptiveCacheTimeoutMs)
-        }
-
-        val cacheEntry = getCacheEntryMetadataIfExists(cacheKeySet.metadataKey)
-
-        attempt = cacheEntry?.attempt?.plus(1) ?: 1
-
-        if (cacheEntry != null && cacheEntry.refreshableAfter.isAfter(Instant.now())) {
-          return resultFromCacheMetadata(cacheEntry, cacheKeySet, typeReference, clazz)
+        webClientCache.tryGetCachedValue(typeReference, requestBuilder, cacheConfig, attempt)?.let {
+          return it
         }
       }
 
@@ -103,54 +64,16 @@ abstract class BaseHMPPSClient(
 
       val result = request.retrieve().toEntity(String::class.java).block()!!
 
-      val deserialized = if (typeReference != null) {
-        objectMapper.readValue(result.body, typeReference)
-      } else {
-        objectMapper.readValue(result.body, clazz)
-      }
+      val deserialized = objectMapper.readValue(result.body, typeReference)
 
       if (cacheConfig != null && requestBuilder.isPreemptiveCall) {
-        val cacheKeySet = getCacheKeySet(requestBuilder, cacheConfig)
-
-        val cacheEntry = PreemptiveCacheMetadata(
-          httpStatus = result.statusCode,
-          refreshableAfter = Instant.now().plusSeconds(cacheConfig.successSoftTtlSeconds.toLong()),
-          method = null,
-          path = null,
-          hasResponseBody = result.body != null,
-          attempt = null,
-        )
-
-        writeToRedis(cacheKeySet, cacheEntry, result.body, cacheConfig.hardTtlSeconds.toLong())
+        webClientCache.cacheSuccessfulWebClientResponse(requestBuilder, cacheConfig, result)
       }
 
       return ClientResult.Success(result.statusCode, deserialized, false)
     } catch (exception: WebClientResponseException) {
       if (cacheConfig != null && requestBuilder.isPreemptiveCall) {
-        val qualifiedKey = getCacheKeySet(requestBuilder, cacheConfig)
-
-        val body = exception.responseBodyAsString
-
-        val backoffSeconds = if (attempt <= cacheConfig.failureSoftTtlBackoffSeconds.size) {
-          cacheConfig.failureSoftTtlBackoffSeconds[attempt - 1].toLong()
-        } else {
-          cacheConfig.failureSoftTtlBackoffSeconds.last().toLong()
-        }
-
-        val cacheEntry = PreemptiveCacheMetadata(
-          httpStatus = exception.statusCode,
-          refreshableAfter = Instant.now().plusSeconds(backoffSeconds),
-          method = method,
-          path = requestBuilder.path ?: "",
-          hasResponseBody = body != null,
-          attempt = attempt,
-        )
-
-        if (attempt >= 4) {
-          Sentry.captureException(RuntimeException("Unable to make upstream request after $attempt attempts", exception))
-        }
-
-        writeToRedis(qualifiedKey, cacheEntry, body, cacheConfig.hardTtlSeconds.toLong())
+        webClientCache.cacheFailedWebClientResponse(requestBuilder, cacheConfig, exception, attempt.get(), method)
       }
 
       if (!exception.statusCode.is2xxSuccessful) {
@@ -163,95 +86,17 @@ abstract class BaseHMPPSClient(
     }
   }
 
-  private fun getCacheKeySet(requestBuilder: HMPPSRequestConfiguration, cacheConfig: PreemptiveCacheConfig): CacheKeySet {
-    val key = requestBuilder.preemptiveCacheKey ?: throw RuntimeException("Must provide a preemptiveCacheKey")
-    return CacheKeySet(preemptiveCacheKeyPrefix, cacheConfig.cacheName, key)
-  }
-
-  private fun writeToRedis(cacheKeySet: CacheKeySet, cacheEntry: PreemptiveCacheMetadata, body: String?, hardTtlSeconds: Long) {
-    redisTemplate.boundValueOps(cacheKeySet.metadataKey).set(
-      objectMapper.writeValueAsString(cacheEntry),
-      Duration.ofSeconds(hardTtlSeconds),
-    )
-
-    if (body != null) {
-      redisTemplate.boundValueOps(cacheKeySet.dataKey).set(
-        body,
-        Duration.ofSeconds(hardTtlSeconds),
-      )
-    }
-  }
-
-  private fun getCacheEntryMetadataIfExists(metaDataKey: String): PreemptiveCacheMetadata? {
-    val stringValue = redisTemplate.boundValueOps(metaDataKey).get()
-      ?: return null
-
-    return objectMapper.readValue<PreemptiveCacheMetadata>(
-      stringValue,
-    )
-  }
-
-  private fun getCacheEntryBody(dataKey: String): String {
-    return redisTemplate.boundValueOps(dataKey).get()
-      ?: throw RuntimeException("No Redis entry exists for $dataKey")
-  }
-
-  private fun <ResponseType> resultFromCacheMetadata(cacheEntry: PreemptiveCacheMetadata, cacheKeySet: CacheKeySet, typeReference: TypeReference<ResponseType>?, clazz: Class<ResponseType>?): ClientResult<ResponseType> {
-    val cachedBody = if (cacheEntry.hasResponseBody) {
-      getCacheEntryBody(cacheKeySet.dataKey)
-    } else {
-      null
-    }
-
-    if (cacheEntry.httpStatus.is2xxSuccessful) {
-      return ClientResult.Success(
-        status = cacheEntry.httpStatus,
-        body = if (typeReference != null) {
-          objectMapper.readValue(cachedBody, typeReference)
-        } else {
-          objectMapper.readValue(cachedBody, clazz)
-        },
-        isPreemptivelyCachedResponse = true,
-      )
-    }
-
-    return ClientResult.Failure.StatusCode(
-      status = cacheEntry.httpStatus,
-      body = cachedBody,
-      method = cacheEntry.method!!,
-      path = cacheEntry.path!!,
-      isPreemptivelyCachedResponse = true,
-    )
-  }
-
   class HMPPSRequestConfiguration {
     internal var path: String? = null
     internal var body: Any? = null
     internal var headers = HttpHeaders()
-    internal var preemptiveCacheConfig: PreemptiveCacheConfig? = null
+    internal var preemptiveCacheConfig: WebClientCache.PreemptiveCacheConfig? = null
     internal var isPreemptiveCall = false
     internal var preemptiveCacheKey: String? = null
     internal var preemptiveCacheTimeoutMs: Int = 10000
 
     fun withHeader(key: String, value: String) = headers.add(key, value)
   }
-
-  data class PreemptiveCacheConfig(
-    val cacheName: String,
-    val successSoftTtlSeconds: Int,
-    val failureSoftTtlBackoffSeconds: List<Int>,
-    val hardTtlSeconds: Int,
-  )
-
-  @JsonInclude(JsonInclude.Include.NON_NULL)
-  data class PreemptiveCacheMetadata(
-    val httpStatus: HttpStatus,
-    val refreshableAfter: Instant,
-    val method: HttpMethod?,
-    val path: String?,
-    val hasResponseBody: Boolean,
-    val attempt: Int?,
-  )
 }
 
 sealed interface ClientResult<ResponseType> {
